[glyphs] Cleanup corner component work to PR
